@@ -1,122 +1,145 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_secure_storage/flutter_secure_storage.dart'; // Import secure storage
+// --- PERBAIKAN: Tambahkan import HomeScreen ---
+import 'package:rhythora/screens/splash_screen.dart';
+import 'package:rhythora/state/auth_cubit.dart'; // Import AuthCubitState
+import 'package:rhythora/state/login_cubit.dart';
+import 'services/auth_service.dart';
+import 'services/player_service.dart';
+import 'services/spotify_service.dart';
+import 'state/player_cubit.dart';
+import 'state/search_cubit.dart';
+
+// --- Konstanta Warna ---
+const Color kBackgroundColor = Color(0xFF121212);
+const Color kCardBackgroundColor = Color(0xFF1F1F1F);
+const Color kCardBackgroundOpacityColor = Color.fromRGBO(39, 39, 42, 0.7);
+const Color kPrimaryColor = Color(0xFF4F46E5); // Indigo-600
+const Color kTextColor = Colors.white;
+const Color kTextSecondaryColor = Color(0xFFA1A1AA); // zinc-400
+const Color kTextMutedColor = Color(0xFF71717A); // zinc-500
+const Color kBorderColor = Color(0xFF3F3F46); // zinc-700
+const Color kFieldFillColor = Color(0xFF27272A); // zinc-800
+// -----------------------
 
 void main() {
+  // Pastikan Flutter binding siap sebelum mengakses service
+  WidgetsFlutterBinding.ensureInitialized();
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // TRY THIS: Try running your application with "flutter run". You'll see
-        // the application has a purple toolbar. Then, without quitting the app,
-        // try changing the seedColor in the colorScheme below to Colors.green
-        // and then invoke "hot reload" (save your changes or press the "hot
-        // reload" button in a Flutter-supported IDE, or press "r" if you used
-        // the command line to start the app).
-        //
-        // Notice that the counter didn't reset back to zero; the application
-        // state is not lost during the reload. To reset the state, use hot
-        // restart instead.
-        //
-        // This works for code too, not just values: Most code changes can be
-        // tested with just a hot reload.
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-      ),
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      _counter++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
-    return Scaffold(
-      appBar: AppBar(
-        // TRY THIS: Try changing the color here to a specific color (to
-        // Colors.amber, perhaps?) and trigger a hot reload to see the AppBar
-        // change color while the other colors stay the same.
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
-      ),
-      body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          //
-          // TRY THIS: Invoke "debug painting" (choose the "Toggle Debug Paint"
-          // action in the IDE, or press "p" in the console), to see the
-          // wireframe for each widget.
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const Text('You have pushed the button this many times:'),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headlineMedium,
+    // Sediakan FlutterSecureStorage agar bisa di-inject ke Service
+    return RepositoryProvider<FlutterSecureStorage>(
+      create: (context) => const FlutterSecureStorage(),
+      child: MultiRepositoryProvider(
+        providers: [
+          RepositoryProvider<AuthService>(
+            // AuthService tidak butuh storage di constructor
+            create: (context) => AuthService(),
+          ),
+          RepositoryProvider<PlayerService>(
+            create: (context) => PlayerService(
+              context.read<AuthService>(),
+            ),
+          ),
+          RepositoryProvider<SpotifyService>(
+             // --- PERBAIKAN 1: Pastikan constructor SpotifyService menerima 2 argumen ---
+             // Jika constructor SpotifyService Anda hanya menerima AuthService, hapus argumen kedua.
+             // Jika constructor butuh keduanya (seperti kode service yang saya berikan), ini sudah benar.
+            create: (context) => SpotifyService(
+              context.read<AuthService>(),
+              context.read<FlutterSecureStorage>(), // SpotifyService butuh storage
+            ),
+             // --------------------------------------------------------------------------
+          ),
+        ],
+        child: MultiBlocProvider(
+          providers: [
+             BlocProvider<AuthCubit>(
+               create: (context) => AuthCubit(
+                 context.read<AuthService>(),
+               // --- PERBAIKAN 2: Nama fungsi salah ---
+               )..checkInitialAuthStatus(), // Ganti nama fungsinya
+               // ------------------------------------
+             ),
+            BlocProvider<SearchCubit>(
+              create: (context) => SearchCubit(
+                context.read<SpotifyService>(),
+              ),
+            ),
+            BlocProvider<PlayerCubit>(
+              create: (context) => PlayerCubit(
+                context.read<PlayerService>(),
+              ),
+            ),
+            BlocProvider<LoginCubit>(
+              create: (context) => LoginCubit(
+                context.read<AuthService>(),
+              ),
             ),
           ],
+          child: MaterialApp(
+            title: 'Rhythora',
+            debugShowCheckedModeBanner: false,
+            theme: ThemeData( // Tema Anda sudah benar
+              brightness: Brightness.dark,
+              scaffoldBackgroundColor: kBackgroundColor,
+              primaryColor: kPrimaryColor,
+              fontFamily: 'Inter', // Pastikan font ada di pubspec.yaml
+              textTheme: const TextTheme(
+                bodyMedium: TextStyle(color: kTextColor),
+              ),
+              textButtonTheme: TextButtonThemeData(
+                style: TextButton.styleFrom(
+                  foregroundColor: kPrimaryColor,
+                  textStyle: const TextStyle(fontWeight: FontWeight.w600),
+                ),
+              ),
+              inputDecorationTheme: InputDecorationTheme(
+                filled: true,
+                fillColor: kFieldFillColor,
+                hintStyle: const TextStyle(color: kTextMutedColor),
+                prefixIconColor: kTextSecondaryColor,
+                suffixIconColor: kTextSecondaryColor,
+                border: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(8.0),
+                  borderSide: const BorderSide(color: kBorderColor),
+                ),
+                enabledBorder: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(8.0),
+                  borderSide: const BorderSide(color: kBorderColor),
+                ),
+                focusedBorder: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(8.0),
+                  borderSide: const BorderSide(color: kPrimaryColor, width: 2.0),
+                ),
+              ),
+              elevatedButtonTheme: ElevatedButtonThemeData(
+                style: ElevatedButton.styleFrom(
+                  backgroundColor: kPrimaryColor,
+                  foregroundColor: Colors.white,
+                  padding: const EdgeInsets.symmetric(vertical: 16.0),
+                  textStyle: const TextStyle(
+                    fontSize: 14.0,
+                    fontWeight: FontWeight.bold,
+                  ),
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(8.0),
+                  ),
+                  elevation: 5.0,
+                ),
+              ),
+            ),
+             home: const SplashScreen(),
+          ),
         ),
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
     );
   }
 }
